Enable cross component analysis for Services & Parameter sources
Summary:
This diff enables taint flow through service IPC which looks in Java code something like

```
public void launcher() {
  Intent cancelNotifIntent = new Intent(context, RTCService.class);
  cancelNotifIntent.setAction(source());
  SecureContext.launchInternalService(cancelNotifIntent, context);
}

public class RTCService extends Service{
  Override
  public int onStartCommand(Nullable Intent intent, int flags, int startId) {
    sink(intent)
  }
}
```

The main difference here with the `getIntent` flow is that `onStartCommand` is an implemented method and also receives the Intent as parameter therefore not need a call-effect. This also enabled many other cases using parameter sources normally like D48431218

Therefore in this diff we apply
- In method_routes_intents_to -> We check the method itself and whether it receives an IPC Intent.
- In get_intent_routing_targets -> Apply either a Root Argument or Root CallEffect depending on whether we are dealing with a call-effect source or a parameter source

Also some refactoring to make this more generically useable for other components we'll need to add.

Things that are still be done (but I deemed out of scope for this diff, feel free to push back)
- We need to at some point migrate these contains over to DSL declaration
- getIntent is still hardcoded but I foresee this being the only calleffect source so I think thats ok for now.

Reviewed By: yuhshin-oss

Differential Revision: D48392861

fbshipit-source-id: fd03655511796e74dededa24a34d97fabaabce2d