Enable cross component analysis for BroadcastReceivers
Summary:
Similar to D48392861 but for BroadcastReceivers. Here I opted for method names instead of signatures since there are many methods that allow you to send broadcasts implemented in many different classes so this keeps things clearer.

A second change in this diff is to make sure that Activity launchers match with Activity Intent handlers, Services with Services and Receivers with Receivers. Otherwise we get False positives like https://fburl.com/code/ydsjczqt where sendBroadcast would map to MainActivity::getIntent which is incorrect.

Reviewed By: yuhshin-oss

Differential Revision: D48466725

fbshipit-source-id: 25b9e5d8f96d6a28e91cb5f852238dcb26d6003e