Remove "enable_class_intervals" flag
Summary:
These feature on/off flags are somewhat painful. We can end up in a state where a feature is on in production, but the E2E tests are running on with the feature off, or vice versa.

Might need a better solution in the long run. For now, I'm removing the flag for class intervals.

Reviewed By: anwesht

Differential Revision: D48256256

fbshipit-source-id: df4264296d832fdc74c275ab2679316e46f32f55