Class intervals: Intersect intervals when creating issues
Summary:
When creating an issue, there should be a non-empty intersection of class intervals between source frames and sink frames. If not, a flow cannot actually occur in practice and the issue would be a false positive.

Intersection between two frames is defined as:
* preserves_type_context = false on either/both frames (i.e. class interval is not applicable) **OR**
* class intervals intersect.

Also, any non-intersecting frames should not be reported as part of the issue.

Reviewed By: arthaud

Differential Revision: D48658545

fbshipit-source-id: 0fb5487500868b5b5a51b547ff3ce29fb8df1101