Remove "field_callee" from Frame
Summary:
Field callee is no longer used since the introduction of `CallInfo::Declaration`. Field callees are found only on declaration frames, which are never emitted outside of debugging purposes. And when they are emitted, the field callee value is the same as the field origins.

The original intention of this field was to show a final hop in the UI to the field itself. This is typically represented by "callee", but that is of type `Method*` which cannot be used for fields.

Diff removes "field callee" altogether.

Reviewed By: arthaud

Differential Revision: D49855343

fbshipit-source-id: 231641bcea4ab3283f5ff490a09804cabdf32a3e