Use patricia trees to implement `FramesMap`
Summary:
We currently use `HashedAbstractPartition` in `FramesMap`, which means the internal maps in the taint representation are using `std::unordered_map` under the hood, which is quite inefficient.
Now that we interned the callee port and represent it with a pointer, all our key types are integer-like types and we can use patricia trees.

Reviewed By: yuhshin-oss

Differential Revision: D49803242

fbshipit-source-id: bc1f34bb413e7ae7f936765bb1c6030b79973d70