Unnest the taint representation
Summary:
# Context
Our taint representation is currently a map of map of map of map, with a very deep level of nesting.
More concretely, it is: Taint: Map[Method, CalleeFrames=Map[Position, CallPositionFrames=Map[CalleePort, CalleePortFrames=Map[Kind, KindFrames=Map[ClassInterval, Frame]]]]].
The benefit of this approach is that each level uses a key which is a pointer type, so we can use `PatriciaTreeMap`.

# Problem
The downside of this approach is the maintainance cost. We have tests for each level (with a lot of duplicates). A single change in that structure requires a lot of effort.

# Solution
This diff proposes to group CalleeFrames, CallPositionFrames, CalleePortFrames into a single structure, LocalTaint. It is indexed by a `CallInfo` which represents the callee, call position, call kind and callee port.
We can use `PatriciaTreeHashMap` under the hood, which allows complex types by implementing a generalized hash map on top of patricia trees (see D49697555).

Reviewed By: anwesht

Differential Revision: D49918986

fbshipit-source-id: 53294adcfb9a7f9dda0e2d50ad88e1f9c7b66257