CRTEX producer: Emit instantiated callees as origins
Summary:
Introduces a new `CrtexOrigin`, which represents the callee + port information at the leaf of a crtex producer.

When canonical names are instantiated, the instantiated name (and port) are also populated in the Frame's origins.

The JSON now emits duplicate information (instantiated canonical name and origins), but it will at least be consistent with the fact that "origins" indicate the "next hop" when call_kind == Origin.

Ideally, we deprecate instantiated canonical names some day too.

**CRTEX Recap:**
* CRTEX producers have user-declared frames/models that provide a canonical name template + "Anchor" as the callee port.
* Canonical name templates are instantiated to concrete canonical names in `KindFrames::propagate()`.
* Instantiated canonical names (and also ports) need to make it into the SAPP trace frame's callee/port, which represents the "next hop" in the flow. They are used by CRTEX consumer issues to connect traces back to the producer.

Reviewed By: arthaud

Differential Revision: D50157000

fbshipit-source-id: 9f82995146460bd1899aeea9df9d02c7b06bbf27