Emit CRTEX producer frames as declarations
Summary:
Before this change, all CRTEX declarations were marked with call kind "origin" instead of "declaration".

However, this is necessary only for the consumer run, which needs to preserve the trace from the declaration back to the producer.

In the producer run, the trace does not include the declaration, just like any non-crtex declaration frame. The declaration contains a templated canonical name, which will be instantiated during propagation, and will be part of the propagated "origin" frame. Therefore, these frames should be created with the "declaration" call kind instead, which is what this diff does.

Reviewed By: anwesht

Differential Revision: D50198132

fbshipit-source-id: 6997522d43c02cd74d061a62e5c5dcd9ea767e9b