CRTEX Consumer: Emit user-defined instantiated canonical names as origins
Summary:
We are trying to achieve the invariant where `if (call kind == origin)` then:
* callee == nullptr AND
* origins indicate where taint comes from:
  * Method/FieldOrigin names the method/field with declared source/sink
  * CrtexOrigin has information that connects the frame/trace to the next hop in a separate run. i.e. where the taint in this run originates from.

In the CRTEX consumer, the origin information is currently defined in the instantiated canonical names. This change copies that information into the "origins" to achieve the invariant above.

TODO: Update parser to read crtex callees from origins (instead of the `normalize_crtex_conditions` hack that we have today).

Reviewed By: anwesht

Differential Revision: D50281334

fbshipit-source-id: 04a75887ffe9a412dc0eec72d5c034fed21e3a00