Always emit "origin" when call info is "Origin"
Summary:
As the title says.

This removes the additional crtex-related checks when determining whether to emit the trace as an "origin". Any frame with call kind == origin will be emitted as such.

Also, the crtex-related checks weren't actually limited to crtex. They also affected parameter sources and return sinks (which has callee port == leaf). With this diff, the "origin" key will also be emitted for those frames (hence the larger-than-expected amount of test changes).

Unfortunately, until the parser is updated to handle `origin > leaves` (next diff), the `call > port` key needs to be emitted as `origin > port`.

Reviewed By: anwesht

Differential Revision: D50202870

fbshipit-source-id: a05cef113202d31673c6c50e045b291db1a0f1a3