Report all of CallInfo as "call_info"
Summary:
Instead of using "call"/"origin" keys based on various conditions, always emit the `CallInfo` directly in a "call_info" key. The parser can then determine whether it's a declaration/origin/callsite based on the call kind and act accordingly.

The keys "call" and "origin" are deprecated after this diff and will be removed once the parser is updated to handle the new "call_info".

Reviewed By: anwesht

Differential Revision: D50818477

fbshipit-source-id: 55f73c7ea3bc3c629e394630fb8df3a37f612ae8