Frame::to_json : Rename "call_info" to "call_kind"
Summary:
Since the parser has been updated to look for "call_kind" inside "call_info", the key for the call kind in `Frame::to_json()` can be renamed.

Note that this key is not read by the parser anymore (that's why we can rename it). Call kind information should be obtained from the call info object, not the frame.

Reviewed By: anwesht

Differential Revision: D51094663

fbshipit-source-id: b890b33642bdaebc8c727d894cbeda07ba742681