Change set_origins_if_declaration to add_origins_if_declaration
Summary:
Previously, `set_origins_if_declaration()` was used to set the origin in the model generator once the method to which the model applied to was known.

However, in C++ model generators, users may define custom origins and we should not override those. (P.S. JSON model generators do not support defining origins.) Change this to `add_origins_if_declaration` so we do not overwrite what has been defined.

Reviewed By: anwesht

Differential Revision: D51224111

fbshipit-source-id: 41fc7828346e4eb7a2edf6a3341ce30e361e8dc8