Highlights: Iterate on LocalTaint instead of Frame
Summary:
Part of the changes for eventually removing `CallInfo` properties from `Frame`. This removes one call-site where we are rely on `CallInfo` properties in `Frame`.

Highlights walk the traces to augment reachable frames with local positions to be highlighted in the UI. `LocalTaint` contains all the next hop information needed for this walk. There is no need to use `visit_frames()`. Use `visit_local_taint()` instead.

Reviewed By: arthaud

Differential Revision: D51481467

fbshipit-source-id: 1ff63c769f8479219ae79a35ec5a458a16da9b5d