Add `CallInfo` argument to `LocalTaint::visit`
Summary:
In a few diffs up the stack, we will be removing `CallInfo`'s properties (callee, callee port, callee position, call kind) from `Frame`. It currently stored in both `LocalTaint` and `Frame`. This requires us to keep their values consistent within the `Taint` structure, which has resulted in bugs and is a waste of memory. The intention is for it to be stored in `LocalTaint` only.

This diff:
* All call-sites that iterate over `Frame`s using `visit_frames()` are updated to accept `CallInfo` as an argument.
* Lambdas will use the `CallInfo` argument for call-info-related properties where it is simple to do so. More complex changes in later diffs.

Reviewed By: arthaud

Differential Revision: D51365584

fbshipit-source-id: 1411fc221e69979acf32cd03dcc2db9ef2a5eab3