LiteralModel: Call `propagate()` at call site
Summary:
Like all other models (`[Method|Field]Model`), `LiteralModel` needs to be "propagated" at the call-site. This ensures that the `call_kind` (and distance and other fields) are set correctly.

Propagating of `LiteralModel` is very similar to that of `FieldModel` and can be shared. Doing this also gives us positions (just the line number) for where the literal model was encountered.

Reviewed By: arthaud

Differential Revision: D51763326

fbshipit-source-id: 4dbf88c1ce9c5e5091162e05bfa32dc385dde7c2