LiteralModel: Set StringOrigin at declaration
Summary:
Reports of hitting the assert at the end of `LocalTaint::update_origin_positions(...)`. This indicates that some origin is not being set.

Turns out that `LiteralModel` does not set origins upon creation. This diff fixes that.

Since literal models describe the pattern that the literal should match, the existing `StringOrigin` is generic enough to be re-used here without having to define a new `LiteralOrigin`.

Reviewed By: arthaud

Differential Revision: D51763325

fbshipit-source-id: 48058b447069bf0a869163c38d6f8d3a55399abe