Only apply strong updates in the backward analysis for side effect propagations
Summary:
See D51982896 for more context.

We do not want to apply strong updates in the backward analysis when the mode `strong-write-on-propagation` is used, since this leads to false positives.
However, we do want strong updates for "side effect" propagations, i.e `Argument(x) -> Argument(x)` propagations.
We actually rely on that behaviour in the `propagation_via_arg` integration test.

Reviewed By: anwesht

Differential Revision: D52036515

fbshipit-source-id: 838376da99efa6cdf73cfaf6c8ccf094c518f889