Update to use d8 instead of dx (when possible)
Summary:
Updating to use `d8` to produce the dex we analyze in integration tests.
- `--core-library`: Removed as no equivalent flag in d8. Looks like it was only suppressing a warning anyway: https://android.googlesource.com/platform/dalvik/+/a9ac3a9d1f8de71bcdc39d1f4827c04a952a0c29/dx/src/com/android/dx/command/dexer/Main.java#89

When testing found that there are some differences in the bytecode generated by dx vs d8. This seems to be a known issue in redex as well. This diff updates to `d8` where we can and optionally allows to select the legacy_dx:
- `field_tito_fn` test case and documents the reasons for the failure.

Reviewed By: yuhshin-oss

Differential Revision: D52095552

fbshipit-source-id: d634e7fac59dfebd21ccdab19d5a010700d112bc