Class Intervals: Narrow the interval on this.* calls
Summary:
When calling `this.someVirtualMethod()`, the receiver's class interval should be that of the most narrow known class for `this`.

Before this change, we were simply using the class interval of the class in which the calling method is defined. Specifically:

```
class CallerClass extends SomeHierarchy {
  void caller() {
    // Class interval of `this` = interval(CallerClass)
    this.someVirtualMethod();
  }
}
```

This works well if `this.someVirtualMethod()` resolves to methods in `CallerClass` or its ancestor/base classes since `CallerClass` is the most-precisely known type for `this`.

However, if it resolves to methods in descendant/derived classes, the type interval needs to be narrowed. This change checks the type interval of the resolved callee's class, and narrows the interval if necessary.

This should address false positives of the form:

```
class Base {
  void issue() {
    s = this.derived1Source();
    this.derived2Sink(s);
  }
}
```

Complicated story: Previously, if `derived1Source()` and `derived2Sink()` reached their respective source/sink via `this.*` calls, we would not find the false positive (correct behavior). The intersection of intervals when the callee's `preserves_type_context=true` resulted in the narrowing of intervals. The false positive happens when the callee's `preserves_type_context=false`, i.e. it's next hop is not a `this.*` call, which requires this change to remove.

Reviewed By: anwesht

Differential Revision: D52142145

fbshipit-source-id: a7a8d72fc713547d59671a69c18062f5814dc8f6