Life-cycles: Call super class method where possible
Summary:
TLDR; This change calls the closest base class method for a life-cycle method, when the method is not overridden in the current class.
___

Previously, when creating a life-cycle method, MT invokes the life-cycle methods (e.g. `onCreate()`, `onStart()`, etc.) only if they are overridden in the current class. This was a design choice made to reduce false positives like the following:

```
class Base {
  void onStart() {
    this.derived2Sink(this.mField);
  }
}

class Derived1 {
  void onCreate() {
    this.mField = source();
  }

  void lifecycle_wrapper() {
    this.onCreate();

    // Resolves to Base;.onStart(), which leads to impossible sink in Derived2
    // But with class intervals, the false positive will not happen.
    // Previously, this was not added to the life-cycle wrapper since Derived1
    // does not define it. With this change, it will be called.
    this.onStart();
  }
}
```

Now that class intervals are supported, such false positives should be impossible.

Doing so however can cause duplicate issues. In the above example, consider the following definition of `Derived2`:

```
class Derived2 {
  void onCreate() {
    this.mField = source();
  }

  void lifecycle_wrapper() {
    // True positive will be found here. **
    this.onCreate();
    this.onStart();
  }
}

// In class Base
  void lifecycle_wrapper() {
    // Resolves to both Derived1 and Derived2.
    // Duplicate issue will be found in Derived2::lifecycle_wrapper (**).
    this.onCreate();
    this.onStart();
  }
```

Next step is to explore the option of creating life-cycle wrappers only in the most-derived class.

Reviewed By: anwesht

Differential Revision: D51994309

fbshipit-source-id: 76a9bf96e9a1032dccec97eda33ca3e1bbff5a87