Create life-cycle methods on most-derived (aka final) classes only
Summary:
Previously, we created life-cycle methods on all classes, as opposed to the most-derived child, because:

* If a method (e.g. `onCreate()`) was not overridden in `Derived`, `Derived.life_cycle_wrapper()` would not call `super.onCreate()`. This causes us to miss taint in the base classes.
  * This is no longer the case.
  * Taint from a base class is reachable via `super.` calls from `Derived.life_cycle_wrapper()`

With that, we should be able to remove life-cycle methods from base classes as the derived class' should see the taint flow if one exists. This also addresses the problem of duplicate issues, one in the base class's life-cycle method and again in the derived class's life-cycle wrapper.

Reviewed By: arthaud

Differential Revision: D51998089

fbshipit-source-id: e87f7f35a37c1d69d0f6d5dc66e72530026a30e3