Optimize the representation of the over and under set abstract domain
Summary:
# Context
The `OverUnderSetAbstractDomain` abstract domain allows to perform both an over-approximation (e.g, represent something that **might** happen) and an under-approximation (represent something that **always** happens).
This is currently implemented with an over set and an under set, with the invariant that the over set includes the under set.

# This diff
To save memory usage, we can avoid storing the same element twice in the over and under set.
This is implemented in this diff by adding the invariant that `over` and `under` are disjoint.
This is trivial for most operations, except the join and the meet.

Reviewed By: yuhshin-oss

Differential Revision: D52659573

fbshipit-source-id: 21d8b4449a8e1d2e20aacac3e7afc8d85a52aff9