Life-cycle+Shims: Shim to derived life-cycle wrapper
Summary:
When shimming to life-cycle wrappers, consider all derived classes.

Life-cycle wrappers are no longer created for non-final classes. If a type resolves to a non-final class, prior to this change, the shim would not be created. This caused false negatives.

The behavior introduced here is similar to the notion of joining models for virtual overrides. However, it is done during call graph construction instead as shims and life-cycles wrappers are not real invoke instructions or methods respectively.

The `kJoinOverrideThreshold` applies too, without which, the analysis times out on bigger apps. If exceeded, the shim would not be created since the non-overridden method is in a non-final class which does not contain any life-cycle wrapper to shim to.

Reviewed By: anwesht

Differential Revision: D53024538

fbshipit-source-id: bf6e234352c8347c9660dbb87a054d00fe1f34b1