Model: Emit source/sink/propagation kind information
Summary:
This will be used to compute rule/category coverage.

A rule is considered covered if for each source/sink kind associated with the rule (and transforms if the rule uses those), there is at least one method that emits that kind.

This diff adds getters for source/sink/transform kinds for the Model that will be used later.

Reviewed By: anwesht

Differential Revision: D53626779

fbshipit-source-id: dd3ff6d7566f29858a40fa84cb643f8ac427b95b