Rule Coverage: Multi-source rules: Both branches must be "covered"
Summary: Multi-source rules can be thought of as 2 single source->sink rules in which both must be triggered. Therefore, they are considered covered only if **both** source->sink rules have corresponding kinds that intersect with the given set of kinds (which are all the known source/sink kinds present in the models at the end of the analysis).

Reviewed By: anwesht

Differential Revision: D53730896

fbshipit-source-id: b2447b66d6111f6e223cc84309ee4a3198646559