Add test for FP shim
Summary: We now only create lifecycle methods for the most derived classes. When creating lifecycle_shims, if the receiver does not consist of the lifecycle wrapper method, we create a shim to all the possible overrides. This diff adds a test to show the potential False Positives of doing so.

Reviewed By: yuhshin-oss

Differential Revision: D53786471

fbshipit-source-id: 9ece78890e8c7091d30d1bc04b1ec875ebfab99e