Add helper method for direct call targets
Summary: This diff additionally distinguishes between static and direct `CallTarget` when constructing the call graph. Lifecycle methods are not virtual but we mimic the "virtual-ness" by shimming to all overrides. But, currently we use `CallTarget::virtual_call()` to set the target for these. This just clarifies these calls and will avoid looking for overrides when computing model_at_callsite.

Reviewed By: yuhshin-oss

Differential Revision: D53825438

fbshipit-source-id: 7676eba6bea1becd8792334507507c4e95527333