Add TypeValue class to track both singleton and small set of DexType for a register
Summary:
This diff stack will use the small set of DexTypes tracked by redex's global types analysis in mariana trench to improve the precision of possible shims and overrides.
This diff adds a new class `TypeValue` to both singleton and small set of DexType for a register. This will be used as the value of the `TypeEnvironment` in the following diff.

Reviewed By: yuhshin-oss

Differential Revision: D53976462

fbshipit-source-id: 69256b1b6fb693edc381b009f6930fb8edecbd48