Use local extends to create lifecycle shims
Summary:
This diff adds a method to access the local extends information for a register. Currently, we create lifecycle shims to all possible derived classes when the precise type is not available. This leads to both FP and FN cases:
- FP: when we create shims to all possible derived types even when only a subset of the derived class is possible.
- FN: when we exceed the max number of overrides (currently set 40) and do not create shims at all.
The `local_extends` types is now a more precise subset of the possible derived types for a base type. When this information is available, we will only create lifecycle shims to this precise subset instead of all possible derived types.

Reviewed By: yuhshin-oss

Differential Revision: D53978382

fbshipit-source-id: 5c2c1d34043789a56d4e60002d5124d21cfcba54