Separate Shim generation and intent routing
Summary: This diff updates the `IntentRouterAnalyzer` to return a unique_ptr which is then moved into the `Shims` for storage instead of using a std::reference_wrapper around a stack object. This also extracts the `IntentRouterAnalyzer` out of `ShimGeneration` as it only deals with user defined shims. This way, we only need to call `IntentRouterAnalyzer::run()` when required.

Reviewed By: arthaud

Differential Revision: D54403293

fbshipit-source-id: 5f8344a4294c328a51cdfffbe6ec8606edec2f7a