Workaround for d8 error on java targets
Summary:
The `io.micrometer` classes are causing errors (`NullPointerException` with v29.0.2, and other compilation errors with v33.0.0) when running `d8`.

Updating our custom desguaring to drop methods in those classes as a workaround for now until we root cause the issue.

Although this gets the build going again, it does result in `d8` emitting a ton of warnings, which we should also get to the root of.

Reviewed By: arthaud

Differential Revision: D55156395

fbshipit-source-id: ec8cc546fafd2b1e465e61b660921044f3cd4577