Fix MethodHandle Desugaring Logic
Summary:
Newer versions of d8 have stricter requirements on the type of the values on the operand stack.

Our desugaring of `MethodHandle.invoke[Exact]` was violating these typing invariants. Newer versions of d8 have stricter requirements on the type of the values on the operand stack.

A `methodHandle.invoke(arg1, ...)` call typically looks like:
```
ALOAD 0 // Push reference to MethodHandle object
// Push arguments
ALOAD 1
...
INVOKEVIRTUAL MethodHandle;.invoke(arg1, ...)[ReturnType]

// Store return value for non-void methods.
[A|I|...]STORE 5
```

**Before**
The current behavior simply removes the `INVOKEVIRTUAL` instruction and pops the arguments.

**After**
With the fix, we still remove the invoke and pop the arguments. However, we also ensure that what is left on the operand stack has a type that matches the subsequent `[A|I|...]STORE` instruction:

* If the return type is an object, the reference to the method handle is on the operand stack and should work with the subsequent `ASTORE` instruction. Nothing else needs to be done.
* If the return type is not an object, the method handle does not need to be on the operand stack because nothing reads it (and d8 will give us stack mismatch warnings). Pop the method handle.
  * For non-void (primitive) return types, the [I|...]STORE method needs to reach something. Push a random byte (BIPUSH) onto the stack so that d8 can handle/type-check the ISTORE instruction without asserting.

Reviewed By: anwesht

Differential Revision: D55264455

fbshipit-source-id: e3403582c77491886136b0c65177f2ac630b1a8e