Rename Field/LiteralModel::from_json to from_config_json
Summary:
Similar to renaming `Model::from_json` to `Model::from_config_json`, this does the rename for `FieldModel` and `LiteralModel`.

The idea is that `to_json()` and `from_json()` should have symmetrical implementations, which is not the case today.

With this change: `to_json()` is used to dump models in the final output, `from_config_json()` is to create models from user-written configs (which has a different schema from the output models).

Reviewed By: arthaud

Differential Revision: D55274854

fbshipit-source-id: 3a9ece7b9ba718fa1902c518887182aa3f6bf416