Adopt DFA AccessScope features
Summary: Deeplinks using DFA are adopting an AccessScope authorization mechanism where no longer the scheme determines if the deeplink handler is accessible but the scope that we set. Therefore abandoning looking at the annotation values but instead look at the `accessScope` value and consider it `PUBLIC` if the accessScope is set to that value. We also check that enforceScope is not `false` otherwise we also consider it to be public.

Reviewed By: anwesht

Differential Revision: D55363707

fbshipit-source-id: 58e548b81613a57a578a31a3ab3f9e122cbf9a63