Add TaggedRoot and TaggedRootSet
Summary:
This structure is intended to add support for tagging via-value and via-type features.

Normally, they are materialized as "via-value:123" or "via-type:int". If the user specifies a tag, it will be "via-<tag>-value:123" or "via-<tag>-type:int".

Instead of using `RootSetAbstractDomain` which only supports specifying the root port which the value/type needs to be materialized from, this structure specifies both root and tag.

Note that the tag is optional and for backward compatibility (and verbosity) reasons, the user specified JSON can be either a string or object, even if that looks a little ugly.

Also, this structure is a `HashedSetAbstractDomain` rather than a `PatriciaTreeSetAbstractDomain` since the `TaggedRoot` pair cannot directly be encoded as an int/pointer. Considering where this is used, it should not have much, if any, noticeable impact on performance.

Reviewed By: anwesht

Differential Revision: D55504309

fbshipit-source-id: dc7c61cb8eb142fcbf8fe19fbf8dc12b79ccc385