Replace AccessPath::is_leaf() with `nullptr`
Summary:
A while back, `AccessPath`s were converted to pointers and created by a factory. This diff further cleans up that path by using pointers for it in `TaintConfig` and `CallInfo`. Similar to `callee`, the `callee_port == nullptr` at leaves/origins.

The motivation behind this is that for serializable classes like `CallInfo`, depending on how the object was created (from `TaintConfig`, or `from_config_json()` or `from_json()`), the leaves/origins have either a callee port of `AccessPath(Root(Root::Kind::Leaf))` or `nullptr`.

This makes checking for equality a pain. I realize we could also override `operator==`, but considering we already have `callee == nullptr` in these cases, it feels cleaner to do the same for `callee_port`.

Ideally, we should not need `Root::Kind::Leaf` and simply use `nullptr`. More changes are needed to achieve that since there are places that operate only on `Root` rather than `AccessPath*`.

Reviewed By: arthaud

Differential Revision: D55614073

fbshipit-source-id: fd41f3ab838caf371489414c03f2bac3f53f6f65