Use matching to_json and from_json implementations for AccessPath components
Summary:
As the title says.

Specifically, this will be used later when parsing propagation output paths which are of type `Path` (rather than `AccessPath` which has a root element on top of the path).

Also created the `to_string()` and `from_string()` equivalents. Like `Kind`, the access path is represented in json as a string rather than a structured object.

Reviewed By: anwesht

Differential Revision: D55347286

fbshipit-source-id: f2c049b01e6049c481863a5b3343fec113c1f0bc