Use local_extends in the CallGraph for more precise override resolution
Summary:
MT uses global type analysis to refine types already when only a single derived type is possible for a given register/variable. This diff further utilizes the results of global type analysis to track a set of possible types for a given register/variable provided as `SmallSetDexTypeDomain`. We then use these types as `local_extends` set (added in D53976477) in `CallTarget::FilterOverrides` during `CallGraph` construction to filter the possible overrides of a virtual call.
This somewhat addresses:
- FN cases when the number of possible overrides exceed the `Heuristics::kJoinOverrideThreshold` currently set to 40.
- FP cases when we can infer the subset of the derived types by not joining method overrides outside of that subset.
**Limitation:** The *maximum* size of the set of possible derived types at a callsite is **4**. After this we will revert back to the previous behavior and can still have both FN and FP mentioned above.

Reviewed By: yuhshin-oss

Differential Revision: D55090524

fbshipit-source-id: 84b2e2afa9b4ae050920aad8ee612104c5b6fdea