Allow non-existing `DexMethod`s to be created
Summary:
Currently, `Model::from_json()` throws an exception when it parses a method that does not exist in the loaded APK.

This is not ideal when parsing sharded models that came from analyzing a dependent .jar which may have methods that the currently loaded APK does not. Instead of throwing an exception, create the underlying `DexMethod*` but mark it with `set_external()` instead.

Reviewed By: anwesht

Differential Revision: D55654518

fbshipit-source-id: ba363ed477a7e1315d0f72732e2c4168e76beb6e