ModelGenerator: Skip parsing .json files. Log warning instead
Summary:
We've moved to using the ".models" extension for model generators long ago. Going to stop parsing ".json" files for models.

Log a warning instead when a ".json" file is skipped. In a properly configured world, the warnings should only show rules/life-cycle/etc. definitions. None of them should be model generators. Going forward, we should use proper extensions for all files to differentiate the file type so that we can remove this warning as well.

Reviewed By: anwesht

Differential Revision: D55940042

fbshipit-source-id: fcf9e2892f3eeddecda6b77ea78b45be09b251a6