TaggedRoot JSON: Always use object format
Summary:
For `TaggedRoot`'s JSON, always use the object format so we don't end up with a mix of object + string in `TaggedRootSet`'s JSON array.

The old format is still supported in `from_json()` (will keep it around for at least a week), but we will print a warning if usage is detected.

Reviewed By: anwesht

Differential Revision: D55936663

fbshipit-source-id: b11a1e91e6956ed6821fd2f86282f35c77007cd4