Relax condition that `DexType` must exist in parameter_type_overrides
Summary:
Seeing a bunch of `DexType` parsing errors when reading in sharded json models:

```
WARNING Unable to parse model `{
... <the model> ...
}`. Expected existing type name for field `type`.
```

Turns out when parsing parameter type overrides, we require the type to exist. With models as input, this may not be the case as these models tend to be generated from a different APK. Relax the condition and create the type if it doesn't exist instead.

Also removing unused `dex_type` and `dex_field` methods in `JsonValidation`.

Reviewed By: anwesht

Differential Revision: D56068175

fbshipit-source-id: 8aadc0d1f426a404f0a1a8853c49ff6d5a78778c