Add sorted partition by kind for consistent iteration order
Summary: Our multi-source/sink rules produce non-deterministic results when both sources are found within the same callable. This is because we iterate on a map with pointer keys (`Kind*`) which does not have a stable value/hash. This diff adds a method `sorted_partition_by_kinds()` which instead creates a vector pairs `(Kind*, Taint)` and sorts it by the string representation of the kind.

Reviewed By: arthaud

Differential Revision: D56152940

fbshipit-source-id: b7a6962b39ff2a779a866759cc6f20a785dd9419