Freeze propagations for sharded models
Summary:
As the title says. This allows sharded models' propagations to override anything user-defined.

Only propagations are frozen because user-defined propagations are usually heuristics, and if we have the actual inferred propagation, those should be preferred.

Sources and sinks should not be frozen because user-defined sources/sinks (including call effect sources/sinks) are not heuristics and should not be overridden regadless.

Reviewed By: anwesht

Differential Revision: D56154024

fbshipit-source-id: 0064fe88019d3669966bf620f5847446c95faab5