Update model generator names for sharded models
Summary:
Add and call `Model::make_sharded_model_generators(identifier)` when reading sharded models. This tells us which models were generated due to the sharded models input vs. user-specified model generators.

It is possible that the sharded models have an empty generators list because everything was inferred. In which case, we add one `ModelGeneratorName` entry to the list to indicated that the model came from a sharded input.

Expanded the `ModelGeneratorName` to take an `is_sharded` flag. Polymorphism is the alternative but feels like overkill for something that can be done with a flag. The identifier for sharded models is the directory name (obtained from the canonical path). This name may be the same as a JSON model generator (which uses a filename), but will be differentiated by the `sharded` flag.

Its string representation will be: "sharded:<directory name>[:<[part = original model generator name>].

If/when we chain cached runs, we can end up with names like "sharded:mydir:sharded:myotherdir:...:original_generator".

Reviewed By: anwesht

Differential Revision: D56218670

fbshipit-source-id: 0434571f7d0a056a26f0d628a35bee60ef9ffd93